Move test case and test suite classes into data package
