+++ core/src/main/kotlin/org/jetbrains/research/testspark/core/test/TestsAssembler.kt
@@ -1,5 +1,7 @@
 package org.jetbrains.research.testspark.core.test
 
+import org.jetbrains.research.testspark.core.test.data.TestSuiteGeneratedByLLM
+
 
 abstract class TestsAssembler {
     private var rawText = ""
