+++ core/src/main/kotlin/org/jetbrains/research/testspark/core/test/data/TestLineType.kt
@@ -1,4 +1,4 @@
-package org.jetbrains.research.testspark.core.test
+package org.jetbrains.research.testspark.core.test.data
 
 /**
  * Represents the types of lines that can be found in a test file.
