+++ core/src/main/kotlin/org/jetbrains/research/testspark/core/test/data/TestSuiteGeneratedByLLM.kt
@@ -1,4 +1,4 @@
-package org.jetbrains.research.testspark.core.test
+package org.jetbrains.research.testspark.core.test.data
 
 /**
  * Represents a test suite generated by LLM.
